pyproject.toml
[tool.poetry]
name = "news-feed"
version = "0.1.0"
description = ""
authors = ["Thomas Cadet <thomas.cadet@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
apache-airflow = {extras = ["google"], version = "^2.9.3"}
apache-airflow-providers-google = "^10.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
